requirements.txt
google-adk
litellm
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
pydantic>=1.8.0
requests>=2.26.0
markdown>=3.3.0
python-multipart>=0.0.5